tests(thesis): Split last file
